+++ Telegram.BotAPI/System/Runtime/CompilerServices/CompilerFeatureRequiredAttribute.cs
@@ -1,11 +1,13 @@
-﻿namespace System.Runtime.CompilerServices;
+namespace System.Runtime.CompilerServices;
 
+#if NETSTANDARD2_0
 /// <summary>Indicates that a compiler feature is required to use the attributed program element.</summary>
 /// <param name="featureName">Name of the required compiler feature.</param>
 [AttributeUsage(AttributeTargets.All, AllowMultiple = true, Inherited = false)]
 #pragma warning disable CS9113 // Параметр не прочитан.
-public sealed class CompilerFeatureRequiredAttribute(string featureName) : Attribute
+internal sealed class CompilerFeatureRequiredAttribute(string featureName) : Attribute
 #pragma warning restore CS9113 // Параметр не прочитан.
 {
     //
 }
+#endif