+++ Telegram.BotAPI/System/Runtime/CompilerServices/IsExternalInit.cs
@@ -1,6 +1,8 @@
-﻿namespace System.Runtime.CompilerServices;
+namespace System.Runtime.CompilerServices;
 
+#if NETSTANDARD2_0
 internal static class IsExternalInit
 {
     //
 }
+#endif