+++ Telegram.BotAPI/System/Runtime/CompilerServices/RequiredMemberAttribute.cs
@@ -1,8 +1,10 @@
-﻿namespace System.Runtime.CompilerServices;
+namespace System.Runtime.CompilerServices;
 
+#if NETSTANDARD2_0
 /// <summary>Indicates that a type or member participates in the required-member feature.</summary>
 [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Field | AttributeTargets.Property, AllowMultiple = false, Inherited = false)]
-public sealed class RequiredMemberAttribute : Attribute
+internal sealed class RequiredMemberAttribute : Attribute
 {
     //
 }
+#endif